[api] More changes for node embedding
Pass in the native `module` and `require` as `__module` and `__require`,
rather than just the native module object. This is because when embedded
the module object is an instance of `NativeModule` (which unlike
`Module` has no require function).

NPM name is now `embedify` so update `embedify.js` to reflect that. Change
the main library from `index.js` to `lib/wrap.js` since we don't need the
rest of the functions in `index.js` (watches, web server, etc).

In `Wrap.prototype.resolver` the `packageFilter` should check for an
`embedify` property not a `browserify` property to override `main` for
embedding (otherwise we might accidentlly source a browser version
somehow has specified).