[fix] Revert behaviour of `require.resolve`
Previously `require.resolve` (like the native version) would throw an exception
if it could not resolve a module. This was changed so we could fall back to
native `require`, but it breaks anything that expects `require.resolve` to behave
in the usual way (e.g. `require.alias`). So, revert that. Instead, in
`require`, trap the exception and fall back as usual.